requirements.txt
flask == 2.2.2
flask-smorest == 0.40.0
python-dotenv == 0.21.0
Flask-SQLAlchemy == 22.3.1